added whiteboard pics and links
